demo-board/scripts/foundry/requirements.txt
azure-identity
azure-ai-projects
azure-ai-agents
